+++ examples/doc_confidence_basic.py
@@ -7,8 +7,7 @@ x = np.linspace(0.3, 10, 100)
 np.random.seed(0)
 y = 1/(0.1*x) + 2 + 0.1*np.random.randn(x.size)
 
-pars = lmfit.Parameters()
-pars.add_many(('a', 0.1), ('b', 1))
+pars = lmfit.create_params(a=0.1, b=1)
 
 
 def residual(p):